requirements.txt
aiohttp==3.8.5
aiortc==1.5.0
av==10.0.0
DateTime==4.7
numpy==1.23.4
opencv-contrib-python==4.7.0.68
opencv-python==4.8.0.76
Pillow==9.3.0
pytest==7.4.2
scikit-learn==1.1.3
scipy==1.9.3


